refactor(shell): extract BackgroundUpdateCoordinator from MainViewModel
MainViewModel was bloated with ~600 lines of background update polling, fallback discovery chains, deduplication tracking, periodic timer management, and one-click update actions. Extracted this entire background subsystem into a dedicated BackgroundUpdateCoordinator service, slashing MainViewModel down to ~270 lines and keeping it strictly focused on shell navigation and view hosting.

Model: gemini-2.5-pro Harness: antigravity
